Modularize the index for better maintainability Replacing "exports.Textarea" by "module.exports" for a easier use
